+++ packages/riverpod/test/framework2/framework_test.dart
@@ -205,6 +205,29 @@ void main() {
     expect(child.read(provider), 42);
   });
 
+  test(
+      'ProviderException not cleared if dependency mayHaveChanged but did not change',
+      () {
+    var callCount = 0;
+    final atom = StateProvider((ref) => 0);
+    final dependency = Provider((ref) => ref.watch(atom).state);
+    final provider = Provider((ref) {
+      callCount++;
+      ref.watch(dependency);
+      if (callCount == 1) {
+        throw Error();
+      }
+    });
+
+    expect(() => container.read(provider), throwsA(isA<ProviderException>()));
+    expect(callCount, 1);
+
+    container.read(atom).state = 0;
+
+    expect(() => container.read(provider), throwsA(isA<ProviderException>()));
+    expect(callCount, 1);
+  });
+
   test('re-evaluating a provider can stop listening to a dependency', () {
     final first = StateProvider((ref) => 0);
     final second = StateProvider((ref) => 0);